python powered raspberry pi
The final file we would run would be a variation of testVision3. This file requires both of the xml files for the implementation of vision of the system. We ran OpenCV  to monitor faces, and we had several trials before we got it to work. There are many dependencies that we do not still have a list of.